Moved "close not modified" button
