+++ htmlContent/git-panel.html
@@ -12,7 +12,6 @@
         </div>
         <div class="btn-group git-available">
             <button title="{{TOOLTIP_REFRESH_PANEL}}" class="btn small git-reset git-icon"><i class="octicon octicon-sync"></i></button>
-            <button title="{{TOOLTIP_CLOSE_NOT_MODIFIED}}"class="btn small git-close-notmodified"><i class="octicon octicon-remove-close"></i></button>
             <button title="{{TOOLTIP_HIDE_UNTRACKED}}" class="btn small git-toggle-untracked"><i class="octicon octicon-eye-unwatch"></i></button>
             <div class="btn-group dropup">
                 <button type="button" class="git-more btn small dropdown-toggle" data-toggle="dropdown" title="{{TOOLTIP_MORE}}">
